+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 - CHANGED: Updated whois.nic.net.sa parser to the new response format.
 
+- CHANGED: Updated whois.nic.cz parser to the new response format.
+
 
 ## Release 3.4.5